Misc improvements and fixes to variant table
